Add files via upload
/*3. Create two classes Named Student and Teacher with required data members. Assume that the
information about the Student and Teacher is stored in a text file. Read n and m number of Student
and Teacher information from the File. Store the information in Array list of type Student and Teacher
Array List<Student> and Array List<Teacher>. Print the information of Teacher who taught OOPS
and Maths. Use Iterator and other functions of util in your program. */