laptop
Declare a class Named Teacher. The class will have all the data members as per your convenient.
The class will have constructors. Write a function to read the values of the class variables. The values
of the variable will be stored in a FILE (text file). The values will be stored in a structured format of
your own choice.