laptop
Watch any of the favorite movie of your choice (any language is fine, preferably English). Create a
Text file to store at least 10 meaningful dialogs from the movie and store it in a text file. Process the
file to remove the stop words (eg. the, is, was, …….) and create another file to have clean text (word).